+++ src/StateSmith/output/StringUtils.cs
@@ -37,6 +37,12 @@ namespace StateSmith.output
             return Regex.Replace(str, @"\r\n|\r|\n", replacment);
         }
 
+        public static string ConvertToSlashNLines(string str)
+        {
+            // TODOLOW compile common regex
+            return Regex.Replace(str, @"\r\n|\r", "\n");
+        }
+
         public static string DeIndentTrim(string str)
         {
             var output = DeIndent(str);
@@ -66,6 +72,25 @@ namespace StateSmith.output
         {
             output = Regex.Replace(output, @"^[ \t]*(\r\n|\r|\n)", "");
             return output;
+        }
+
+        internal static string EscapeCharsForString(string str)
+        {
+            if (str == null) return null;
+
+            str = Regex.Replace(str, @"(\\|"")", "\\$1");
+            return str;
+        }
+
+        internal static string RemoveEverythingBeforeRequiredMatch(string str, string match)
+        {
+            var index = str.LastIndexOf(match);
+            if (index < 0)
+            {
+                throw new ArgumentException($"match `{match}` not found");
+            }
+
+            return str.Substring(index + match.Length);
         }
     }
 }